requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
sqlalchemy==2.0.41
pandas==2.2.2
passlib==1.7.4
python-jose==3.5.0
python-multipart==0.0.20
pydantic==2.11.5
bcrypt==3.2.2
matplotlib==3.9.0
seaborn==0.13.2
reportlab==4.4.1
Pillow==10.3.0
scikit-learn==1.7.0
scipy==1.15.3
psycopg2-binary==2.9.9 